Rename YaruCompactLayout to YaruNavigationPage
Close: #416
