+++ test/pages/goldens/yaru_navigation_page-compact-dark.png
