+++ test/pages/goldens/yaru_navigation_page-compact-light.png
