+++ test/pages/goldens/yaru_navigation_page-labelled-dark.png
