+++ test/pages/goldens/yaru_navigation_page-labelled-extended-dark.png
